TA-175: rename and repackage ccd consumer tests
