+++ src/contractTest/java/uk/gov/hmcts/probate/service/consumer/ccd/StartForCaseworkerConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.service.consumer;
+package uk.gov.hmcts.probate.service.consumer.ccd;
 
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactHttpsProviderRuleMk2;
@@ -23,7 +23,7 @@ import static uk.gov.hmcts.reform.probate.pact.dsl.PactDslBuilderForCaseDetailsL
 @SpringBootTest({
     "core_case_data.api.url: localhost:4456"
 })
-public class ProbateBackOfficeStartForCaseworker extends AbstractBackOfficePact {
+public class StartForCaseworkerConsumerTest extends AbstractCcdConsumerTest {
 
     @Rule
     public PactHttpsProviderRuleMk2 provider = new PactHttpsProviderRuleMk2("ccdDataStoreAPI_Cases", "localhost", 4456, this);
